feat: implement user authentication and authorization, update restaurant and review endpoints to require current user context

+++ app/services/restaurant.py
@@ -3,20 +3,20 @@ import models.restaurants, models.reviews, models.users
 from fastapi import HTTPException, status
 
 
-def create_restaurant(db, restaurant):
+def create_restaurant(db, restaurant, current_user):
     result = db.execute(
         select(models.restaurants.Restaurants).where(models.restaurants.Restaurants.title == restaurant.title)
     )
     existing_restaurant = result.scalars().first()
     if existing_restaurant:
         raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Restaurant already exists")
-
+    
     new_restaurant = models.restaurants.Restaurants(
         title=restaurant.title,
         location=restaurant.location,
         cuisine=restaurant.cuisine,
         contact_number=restaurant.contact_number,
-        owner_id=1
+        owner_id=current_user.id
         
 
     )
@@ -25,7 +25,7 @@ def create_restaurant(db, restaurant):
     db.refresh(new_restaurant)
     return new_restaurant
 
-def get_restaurant(restaurant_id, db):
+def get_restaurant(db, restaurant_id):
     result = db.execute(select(models.restaurants.Restaurants).where(models.restaurants.Restaurants.id == restaurant_id))
 
     existing_restaurant = result.scalars().first()
@@ -46,22 +46,28 @@ def get_all_restaurants_by_cuisine(cuisine, db):
         raise HTTPException(404, "No restaurants found")
     return restaurants
 
-def delete_restaurant(restaurant_id, db,):
+def delete_restaurant(restaurant_id, db,current_user):
     result = db.execute(select(models.restaurants.Restaurants).where(models.restaurants.Restaurants.id == restaurant_id))
     existing_restaurant = result.scalars().first()
     if not existing_restaurant:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Restaurant not found")
+
+    if existing_restaurant.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="You are not authorized to delete this restaurant")
     
     db.delete(existing_restaurant)
     db.commit()
     return 
 
 
-def update_restaurant(restaurant_id, db, restaurant):
+def update_restaurant(restaurant_id, db, restaurant, current_user):
     result = db.execute(select(models.restaurants.Restaurants).where(models.restaurants.Restaurants.id == restaurant_id))
     existing_restaurant = result.scalars().first()
     if not existing_restaurant:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Restaurant not found")
+
+    if restaurant.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="You are not authorized to update this restaurant")
     
     if restaurant.title is not None:
         existing_restaurant.title = restaurant.title
